Synchronize all projects on refresh action
This reverts commit 53279b7f15336159b40f7a470daf2fa6603000c9.

After further discussion we decided not to do "partial refreshes", as that would
leave the workspace in an inconsistent state. It would also mean that closing
and reopening Eclipse leads to different configuration, because opening Eclipse
leads to a full refresh anyway.

If refreshing all projects turns out to be a performance bottleneck for some users,
we could allow manual definition of composites.

+review BUILDSHIP-10