Enable cyclic project renaming
We now handle cyclic project renamings as long as the two projects are
part of the same composite (which is currently true for all Gradle
projects in the workspace). If A and B are about to swap their names, we
can safely assign a temporary name to B. If only A is to be renamed into
B, but B keeps its name, then we throw an Exception as before.