Wait for jobs to finish in test cleanup
Deleting projects now triggers a synchronization job, which needs to be
allowed to finish before we can continue testing. Otherwise it could end
up removing project configuration for projects created in the next test.

The wait implementation was improved to join on a common job category
instead of doing an idle loop.