Don't reimport projects that were just deleted
Up until now, the project synchronization would always import all
subprojects that it found in a build. This breaks when we try to delete
a project. The deletion triggers a refresh operation and the refresh
operation re-creates the project.

The synchronization now accepts a strategy for what to do with projects
that are not yet in the workspace. This strategy also replaces the
`workingsets` argument with a more general `afterImport` callback. The
import wizard can use this callback to add the workingsets directly
instead of relying on an event mechanism.

There also was a minor inconsistency in how Java source settings were
handled. Some code paths set them eagerly while other code paths relied
on the fact that the classpath container updater would set them later. I
made these code paths set the source settings consistently, so the tests
don't have to wait for the extra job.