Remove single-project synchronization methods from API
There was no more place in the code base where single projects were
synchronized. Synchronizing just a single project inside a composite
would leave the workspace in an inconsistent state.

There is actually still potential inconsistency here, as we don't
atomically synchronize the whole composite, but one multi-project-build
at a time. The only reason for this is that we need to pass along the
FixedRequestAttributes. We should get rid of this and change the API to
accept an OmniEclipseWorkspace instead. Until then we need to be careful
to use this API correctly.