Synchronize the whole OmniEclipseWorkspace consistently
This changes the WorkspaceGradleOperations to take the whole
OmniEclipseWorkspace for synchronization instead of individual Gradle
builds. Only by doing this will the Eclipse workspace be in a consistent
state afterwards (all renamings consistently applied, dependency
substitution consistently applied etc.)

This also exposes how inconvenient it currently is to correlate the
OmniXXX models back to the build (FixedRequestAttributes) that they came
from.